refactor: Update imports and variable names
